refactor: overhaul signs fetch task
