+++ src/navigation/TabNavigation.tsx
@@ -15,7 +15,7 @@ import Messages from '../screens/messages/Messages';
 import AboutSignsDetails from '../screens/signs/AboutSignsDetails';
 import TimeTablePage from '../screens/timeTable/TimeTable';
 import { registerReminderTask } from '../tasks/disciplineTasks';
-import { registerSignsFetchTask } from '../tasks/signs';
+import { registerSignsFetchTask } from '../tasks/signs/signs';
 import { AppShortcutItem } from '../utils/shortcuts';
 import ServicesStackNavigator from './ServicesStackNavigator';
 import SignsTopTabNavigator from './TopTabNavigator';
