+++ src/screens/settings/ToggleSignNotification.tsx
@@ -6,7 +6,7 @@ import { cache } from '../../cache/smartCache';
 import Text from '../../components/Text';
 import { useAppDispatch, useAppSelector, useGlobalStyles } from '../../hooks';
 import { setSignNotification } from '../../redux/reducers/settingsSlice';
-import { registerSignsFetchTask, unregisterBackgroundFetchAsync } from '../../tasks/signs';
+import { registerSignsFetchTask, unregisterBackgroundFetchAsync } from '../../tasks/signs/signs';
 import { NOTIFICATION_GUIDE_URL } from '../../utils';
 import { fontSize } from '../../utils/texts';
