+++ src/tasks/signs/types.ts
@@ -1,0 +1,7 @@
+import { ICheckPoint } from '../../models/sessionPoints';
+
+export interface IDifferentCheckPoint {
+  oldResult: ICheckPoint;
+  newResult: ICheckPoint;
+  subjectName: string;
+}
